optimize code analysis service and improve logging
- Remove unused methods from CodeAnalysisService and AIServiceFactory
- Simplify UserMessage model by removing redundant code field
- Make enhanceRequestWithMemory strategy-agnostic and consistent
- Move LLM prompt logging to correct location (after strategy processing)
- Clean up duplicate system prompts and special handling
- Improve code consistency and maintainability

Changes:
- CodeAnalysisService: Remove unused analyzeCode() and streamAnalysis() methods
- AIServiceFactory: Remove unused getBestAvailableService() method
- UserMessage: Remove redundant code field, simplify message creation
- enhanceRequestWithMemory: Make uniform for all strategies, remove special cases
- AI Services: Add proper LLM prompt logging after strategy processing
- Remove duplicate system prompts from service layer
- Consistent code block formatting for all analysis types

This improves code quality, reduces redundancy, and provides accurate logging
of the actual prompts sent to LLM after strategy modifications.
