+++ executor/src/main/resources/templates/landing.html
@@ -193,76 +193,62 @@
         <div class="services-grid">
             <!-- Code Assistant Service -->
             <div class="service-card">
-                <span class="service-icon">💻</span>
+                <span class="service-icon">🧠</span>
                 <h2 class="service-title">Code Assistant</h2>
                 <p class="service-description">
-                    Intelligent code analysis and assistance powered by advanced AI models. 
-                    Get explanations, refactoring suggestions, debugging help, and comprehensive code analysis.
+                    Advanced AI-powered code assistant with streaming, memory, and tool capabilities. 
+                    Get intelligent explanations, refactoring suggestions, and comprehensive analysis with persistent context.
                 </p>
                 <ul class="service-features">
-                    <li>Code explanation and documentation</li>
-                    <li>Refactoring suggestions and improvements</li>
-                    <li>Bug detection and debugging tips</li>
-                    <li>Comprehensive code analysis</li>
+                    <li>Conversational AI with memory and follow-up questions</li>
+                    <li>Real-time streaming responses</li>
+                    <li>Advanced code analysis and refactoring</li>
+                    <li>Session management with conversation history</li>
+                    <li>Markdown formatting with syntax highlighting</li>
                     <li>Support for 14+ programming languages</li>
                     <li>Multiple AI providers (OpenAI GPT, Llama)</li>
+                    <li>Context-aware RAG layer (coming soon)</li>
+                    <li>LangChain4j agent support with tools (coming soon)</li>
                 </ul>
                 <a href="/code-assistant" class="service-button">Try Code Assistant</a>
             </div>
 
-            <!-- AI Chat Service -->
-            <div class="service-card">
-                <span class="service-icon">💬</span>
-                <h2 class="service-title">AI Chat Assistant</h2>
-                <p class="service-description">
-                    Conversational AI assistant for technical discussions, problem-solving, 
-                    and knowledge sharing. Perfect for developers and technical teams.
-                </p>
-                <ul class="service-features">
-                    <li>Natural language conversations</li>
-                    <li>Technical problem solving</li>
-                    <li>Knowledge sharing and learning</li>
-                    <li>Context-aware responses</li>
-                    <li>Multi-turn conversations</li>
-                    <li>Integration with development workflows</li>
-                </ul>
-                <a href="/ai-chat" class="service-button">Try AI Chat</a>
-            </div>
-
             <!-- Code Generator Service -->
-            <div class="service-card">
-                <span class="service-icon">⚡</span>
-                <h2 class="service-title">Code Generator</h2>
+            <div class="service-card coming-soon">
+                <span class="service-icon">📄</span>
+                <h2 class="service-title">Document RAG</h2>
                 <p class="service-description">
-                    Generate production-ready code from natural language descriptions. 
-                    Create APIs, functions, classes, and complete applications with AI assistance.
+                    Upload documents (PDF, text) and ask questions using RAG. 
+                    Get intelligent answers from your documents with vector search and chat memory.
                 </p>
                 <ul class="service-features">
-                    <li>Natural language to code conversion</li>
-                    <li>API generation from specifications</li>
-                    <li>Database schema generation</li>
-                    <li>Test case generation</li>
-                    <li>Multiple framework support</li>
-                    <li>Best practices integration</li>
+                    <li>File upload UI for PDF and text documents</li>
+                    <li>LangChain4j file loaders and chunking</li>
+                    <li>Vector DB storage with embeddings</li>
+                    <li>Semantic search and retrieval</li>
+                    <li>Multi-turn chat history</li>
+                    <li>Support for multiple documents per session</li>
+                    <li>Hybrid prompting with context chunks</li>
                 </ul>
-                <a href="/code-generator" class="service-button">Try Code Generator</a>
+                <a href="#" class="service-button">Coming Soon</a>
             </div>
 
             <!-- Documentation Generator (Coming Soon) -->
             <div class="service-card coming-soon">
-                <span class="service-icon">📚</span>
-                <h2 class="service-title">Documentation Generator</h2>
+                <span class="service-icon">🛫</span>
+                <h2 class="service-title">Flight Deal Bot</h2>
                 <p class="service-description">
-                    Automatically generate comprehensive documentation from your codebase. 
-                    Create API docs, README files, and technical documentation with AI.
+                    Smart assistant that helps users find best flight deals using LLM and RAG. 
+                    Get intelligent flight recommendations with price analysis and trends using mock data.
                 </p>
                 <ul class="service-features">
-                    <li>Automatic API documentation</li>
-                    <li>README file generation</li>
-                    <li>Code comment enhancement</li>
-                    <li>Technical specification docs</li>
-                    <li>Multiple format support</li>
-                    <li>Version control integration</li>
+                    <li>Mock flight data with realistic city pairs and prices</li>
+                    <li>RAG-powered deal search and recommendations</li>
+                    <li>Price analysis and trend detection</li>
+                    <li>Advanced filtering (budget, dates, airlines)</li>
+                    <li>LLM-powered deal ranking and summarization</li>
+                    <li>Chat interface with session memory</li>
+                    <li>Weather and visa requirement tools (optional)</li>
                 </ul>
                 <a href="#" class="service-button">Coming Soon</a>
             </div>
